Allow 1-D input where appropriate
For MLR and CLS it is reasonable to want to use vectors for both X and
Y calibration data. For PCR and PLS1 it is reasonable to use a vector
for the Y calibration data if only one output is needed. The PLS2
algorithm is the same as the PLS1 algorithm if only one Y variable is
required, so there is no need for PLS2 to accept vector input.

In each case, the vector is reshaped to a (?,1) matrix - essentially a
column vector.