PRESS and Q2 now take the number of CV groups
Previously the cross-validation was done using the miss-one-out
technique, but that has two drawbacks - it can be slow and it can
produce some odd statistical artefacts. These artefacts arise on those
cross-validation iterations where the values being missed-out are
relative outliers. The training set on these iterations will not be
representative of the whole calibration data.

By dividing the data set into groups and missing one out each time the
total number of iterations needed can be reduced significantly, and each
iteration should be working with a more representative training set.

The old behaviour can be achieved by setting the number of groups to be
equal to the number of data samples.